made binocssoftware file with gui as option
