+++ binocs/python/binocs/io.py
@@ -94,6 +94,7 @@ def readdataframe(options: dict) -> Tuple[pd.DataFrame, pd.DataFrame]:
 		"ra" : options_df["ra"],
 		"dec": options_df["dec"],
 		# "rv" : options_df["rv"]
+		"rv": 0
 	})
 
 	mag_df = pd.DataFrame({
@@ -133,7 +134,8 @@ def readdataframe(options: dict) -> Tuple[pd.DataFrame, pd.DataFrame]:
         'B4_err' : options_df['B4_err'],
 	})
 
-	mag_df = mag_df.fillna(99.99999999999999)
+	mag_df = mag_df.fillna(99.999)
+	info_df = info_df.fillna(99.999)
 
 	print("    %d stars in file." % (len(info_df)))
 	return info_df.to_numpy(), mag_df.to_numpy()
